admin delete all users + view/delete all placemarks + menu adjustments

+++ app/src/main/java/org/wit/property_manager/activities/PropertyListActivity.kt
@@ -80,10 +80,14 @@ class PropertyListActivity : AppCompatActivity(), PropertyListener {
         }
         when (item.itemId) {
             R.id.item_deleteAll -> {
-                app.properties.deleteByUser(currentUser.id)
-                val launcherIntent = Intent(this, PropertyListActivity::class.java)
-                launcherIntent.putExtra("current_user", currentUser).putExtra("user", currentUser)
-                startActivityForResult(launcherIntent,0)
+                if (isAdmin){ app.properties.deleteAll()}
+                else {
+                    app.properties.deleteByUser(currentUser.id)
+                    val launcherIntent = Intent(this, PropertyListActivity::class.java)
+                    launcherIntent.putExtra("current_user", currentUser)
+                        .putExtra("user", currentUser)
+                    startActivityForResult(launcherIntent, 0)
+                }
             }
         }
         return super.onOptionsItemSelected(item)
